Add task for "Refactor tasks tracker" due to a change in the design.
We came into conclusion that the PromptStrategy will need more context than just the task itself.
Therefore we also expose the tracker document's body and optionally the completed tasks to it.
I had to reset the task of PromptStrategy and add this refactor, after the refactor we'll do it again